requirements.txt
pydantic
requests
python-jose
azure-functions
